Cleanup

+++ src/dex/BlasterSwapV2Vault.sol
@@ -93,8 +93,8 @@ contract BlasterSwapV2Vault is BaseDexUniformVault {
         (uint112 reserve0, uint112 reserve1,) = lpToken.getReserves();
         uint160 sqrtPrice = getCurrentSqrtPrice();
         return isToken0
-            ? reserve0 + Math.mulDiv(reserve1, 2 ** 192, sqrtPrice * sqrtPrice)
-            : reserve1 + Math.mulDiv(reserve0, sqrtPrice * sqrtPrice, 2 ** 192);
+            ? reserve0 + Math.mulDiv(reserve1, 2 ** 192, uint256(sqrtPrice) * uint256(sqrtPrice))
+            : reserve1 + Math.mulDiv(reserve0, uint256(sqrtPrice) * uint256(sqrtPrice), 2 ** 192);
     }
 
     /* ========== INTERNAL FUNCTIONS ========== */
